+++ CMakeLists.txt
@@ -17,7 +17,7 @@ add_executable(raytracer
         Source/triangle.h
         Source/utils.cpp
         Source/utils.h
-        Source/vector_type.h Source/math_utils.h Source/ray.h Source/camera.h Source/object.h Source/intersection.h Source/material.h Source/geometry.h Source/scene.h Source/light.h Source/surfaceinteraction.h Source/sampler.h Source/texture.h Source/math_vector_utils.h Source/bsdf_lobe.h Source/boundingbox.h)
+        Source/vector_type.h Source/math_utils.h Source/ray.h Source/camera.h Source/object.h Source/intersection.h Source/material.h Source/geometry.h Source/scene.h Source/light.h Source/surfaceinteraction.h Source/sampler.h Source/texture.h Source/math_vector_utils.h Source/bsdf_lobe.h Source/boundingbox.h Source/pathtrace.h)
 
 find_library(SDL_LIB libsdl2 HINTS include/sdl/lib)
 target_link_libraries(raytracer PUBLIC ${SDL_LIB})
