+++ Source/main.cpp
@@ -1,6 +1,7 @@
 #include "camera.h"
 #include "light.h"
 #include "object.h"
+#include "pathtrace.h"
 #include "ray.h"
 #include "raytrace.h"
 #include "sampler.h"
@@ -40,7 +41,7 @@ scg::Camera camera{
     SCREEN_WIDTH,
     SCREEN_HEIGHT,
     true,
-    0.05f,
+    0.2f,
     3.0f};
 
 scg::Scene scene;
@@ -88,8 +89,8 @@ void Draw(screen *screen)
             scg::Ray ray = camera.getLensRay(x, y, sampler[omp_get_thread_num()]);
             ray.minT = scg::RAY_EPS;
 
-            int depth = 4;
-            float gamma = 1.0f;
+            int depth = 10;
+            float gamma = 3.0f;
             scg::Vec3f colour = scg::trace(scene, ray, depth, sampler[omp_get_thread_num()]);
             buffer[y][x] += colour * gamma; // TODO: clamp value
